--- Assets/Puppet/PuppetOther.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 8b89374df75542d44a50b8fcea440c8d
-timeCreated: 1501075373
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
